Checkout
Changed checkout index endpoint to recieve the restaurant and return to the menu of that same restaurant if not possible to checkout.

Changed Index View from checkout to send to the views the list of payment methods.

Changed endpoint to finish the checkout to receive the PaymentMethod selected and verify if is valid to finish the checkout.